Remove subscription context and fetch status on demand
